refactor: remove duplicate list rendering
